requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.10
pandas>=2.2.2
pandas-ta==0.4.71b0
numpy>=2.2.6
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.9.0
httpx>=0.27.0
pytest>=8.2.0
hypothesis>=6.100.1
pytest-asyncio>=0.23.6
